Uniformed SoA with AoS version
