--- SoA/imgs/output/.gitignore
@@ -1,4 +1,0 @@
-# Ignore every file in this folder
-*
-# Except this one
-!.gitignore
