--- SoA/src/KernelFactory.h
@@ -1,19 +1,0 @@
-#ifndef KERNELFACTORY_H
-#define KERNELFACTORY_H
-#include <memory>
-
-#include "Kernel.h"
-
-
-class KernelFactory {
-public:
-    KernelFactory() = delete;
-    ~KernelFactory() = delete;
-
-    static std::unique_ptr<Kernel> createBoxBlurKernel(unsigned int order);
-    static std::unique_ptr<Kernel> createEdgeDetectionKernel(unsigned int order);
-};
-
-
-
-#endif //KERNELFACTORY_H
