+++ SoA/src/image/Image.cpp
