+++ SoA/src/image/reader/ImageReader.cpp
