+++ SoA/src/image/reader/STBImageReader.cpp
