+++ SoA/src/kernel/KernelFactory.cpp
