+++ SoA/src/processing/ImageProcessing.cpp
