+++ SoA/tests/CMakeLists.txt
@@ -18,4 +18,7 @@ add_executable(kip_sequential_runTests ${TEST_SOURCES})
 
 target_link_libraries(kip_sequential_runTests kip_sequential_lib gtest_main gmock_main)
 
-add_test(NAME runTests COMMAND runTests)
+add_compile_definitions(TEST_IMAGES_INPUT_DIRPATH="${PROJECT_SOURCE_DIR}/tests/imgs/input/")
+add_compile_definitions(TEST_IMAGES_OUTPUT_DIRPATH="${PROJECT_SOURCE_DIR}/tests/imgs/output/")
+
+add_test(NAME runAllTests COMMAND kip_sequential_runTests)
